+++ app/Models/MataKuliah.php
@@ -13,7 +13,6 @@ class MataKuliah extends Model
     protected $fillable = [
         'kode_mk',
         'nama_mk',
-        'dosen_id',
         'sks',
         't_p',
         'semester'
@@ -22,11 +21,5 @@ class MataKuliah extends Model
     public function jadwal()
     {
         return $this->hasMany(Jadwal::class);
-    }
-
-    // Reverse one to one dosen ke mata kuliah
-    public function dosen()
-    {
-        return $this->belongsTo(Dosen::class)->withDefault();
     }
 }
