+++ resources/views/dashboard/jadwal/edit.blade.php
@@ -44,6 +44,17 @@
                                 @endforeach
                             </select>
                         </div>
+                        <div class="input-group input-group-static mb-3 @error('dosen') is-invalid @enderror is-filled">
+                            <label for="exampleFormControlSelect2" class="ms-0">Dosen</label>
+                            <select class="form-control" id="exampleFormControlSelect2" name="dosen_id">
+                                @foreach ($dosen as $row)
+                                <option value="{{ $row->id }}" {{ ($row->id == $jadwal->dosen_id) ?
+                                    'selected' :
+                                    ''
+                                    }} >{{ $row->nama_dosen }}</option>
+                                @endforeach
+                            </select>
+                        </div>
                         <div
                             class="input-group input-group-static my-3 @error('kelas') is-invalid is-filled @enderror is-filled">
                             <label class="form-label">Kelas</label>
