+++ resources/views/dashboard/mata_kuliah/index.blade.php
@@ -37,9 +37,6 @@
                                         Kode MK</th>
                                     <th
                                         class="bg-gradient-info text-white text-uppercase text-secondary text-xxs font-weight-bolder ps-2">
-                                        Nama Mata Kuliah</th>
-                                    <th
-                                        class="bg-gradient-info text-white text-uppercase text-secondary text-xxs font-weight-bolder ps-2">
                                         Dosen</th>
                                     <th
                                         class="bg-gradient-info text-white text-uppercase text-secondary text-xxs font-weight-bolder ps-2">
@@ -66,11 +63,6 @@
                                     <td class="align-middle">
                                         <p class="text-xs font-weight-bold mb-0">
                                             {{ $row->nama_mk }}
-                                        </p>
-                                    </td>
-                                    <td class="align-middle">
-                                        <p class="text-xs font-weight-bold mb-0">
-                                            {{ $row->dosen->nama_dosen }}
                                         </p>
                                     </td>
                                     <td class="align-middle">
