+++ resources/views/view.blade.php
@@ -141,7 +141,7 @@
                                                 </td>
                                                 <td class="align-middle">
                                                     <p class="text-sm font-weight-bold mb-0">
-                                                        {{ $row->jadwal->mataKuliah->dosen->nama_dosen }}
+                                                        {{ $row->jadwal->dosen->nama_dosen }}
                                                     </p>
                                                 </td>
                                                 <td
